move all what applies to ajax requests to api file
